Partial type, PropsWithChildren
