+++ src/components/v5/common/ActionSidebarRow/types.ts
@@ -7,7 +7,6 @@ export interface ActionSidebarRowProps {
   isDescriptionFieldRow?: boolean;
   isOpened?: boolean;
   onToggle?: () => void;
-  children?: React.ReactNode;
   isError?: boolean;
   fieldName: ActionSidebarRowFieldNameEnum;
 }
