+++ src/components/v5/shared/Stepper/partials/StepperButton/consts.ts
@@ -1,8 +1,12 @@
 import { StepStage } from './types';
 
-export const iconNameMap: Record<StepStage, string | undefined> = {
-  completed: 'check',
-  current: undefined,
-  upcoming: undefined,
-  skipped: undefined,
+export const STEP_STAGE = {
+  Completed: 'completed',
+  Current: 'current',
+  Upcoming: 'upcoming',
+  Skipped: 'skipped',
+} as const;
+
+export const iconNameMap: Partial<Record<StepStage, string | undefined>> = {
+  [STEP_STAGE.Completed]: 'check',
 };
